requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
importlib-metadata==6.7.0
outcome==1.2.0
packaging==23.1
PySocks==1.7.1
python-dotenv==0.21.1
requests==2.31.0
selenium==4.11.2
sniffio==1.3.0
sodapy==2.2.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.7.1
urllib3==1.26.6
webdriver-manager==4.0.0
wsproto==1.2.0
zipp==3.15.0
